+++ src/component.spec.ts
@@ -3,7 +3,6 @@ import {
   action,
   props,
   setGroup,
-  stateOf,
   spaceOf,
   sendMsg,
   toChild,
@@ -190,46 +189,6 @@ describe('Component helpers', () => {
     it ('toChild should send an undefined message to a child component from the parent correctly', () => {
       toChild(app.ctx.components['MyComp'].ctx, 'Child', 'childInput')
       expect(childData).toEqual(undefined)
-    })
-
-  })
-
-  describe('stateOf helper', () => {
-    let lastError
-    let ctx = {
-      id: 'id1',
-      components: {
-        id1: {
-          state: {},
-        },
-        id1$child: {
-          state: {},
-        },
-      },
-      error: (source, description) => lastError = [source, description],
-    }
-    let ctx2 = {
-      ...ctx,
-      id: 'wrong',
-    }
-
-    it('should return the component state from her context', () => {
-      expect(stateOf(<any> ctx)).toBe(ctx.components[ctx.id].state)
-    })
-
-    it('should get the state from a certain component', () => {
-      let state = stateOf(<any> ctx , 'child')
-      expect(state).toEqual(ctx.components['id1$child'].state)
-    })
-
-    it('should log an error if there are no ctx space', () => {
-      stateOf(<any> ctx2)
-      expect(lastError).toEqual(['stateOf', `there are no space 'wrong'`])
-    })
-
-    it('should log an error if there are no child space', () => {
-      stateOf(<any> ctx, 'wrongChild')
-      expect(lastError).toEqual(['stateOf', `there are no child 'wrongChild' in space 'id1'`])
     })
 
   })